Check for file existence in Tempfile#unlink
This removes the need to reset the temp file instance variable in the
tests (which wasn't required for MRI)

+++ lib/tempfile.rb
@@ -51,7 +51,7 @@ class Tempfile
   alias length size
 
   def unlink
-    File.unlink(@tmpfile.path)
+    File.unlink(@tmpfile.path) if File.exist?(@tmpfile.path)
   end
   alias delete unlink
